fix: webhook_limit BIGINT type mismatch — i32 → i64
Migration 046 changed webhook_limit from INT to BIGINT but the
Customer struct and related types still used i32. This caused
sqlx deserialization failure on all customer queries.

Same root cause as the webhook_count fix (migration 011). Both
columns are now BIGINT in the database and i64 in Rust.

+++ api/src/routes/admin.rs
@@ -106,7 +106,7 @@ pub struct UserDetail {
     pub plan: String,
     pub is_active: bool,
     pub is_admin: bool,
-    pub webhook_limit: i32,
+    pub webhook_limit: i64,
     pub webhook_count: i64,
     pub created_at: chrono::DateTime<chrono::Utc>,
     pub endpoints: Vec<EndpointSummary>,
